test: add custom fee tests with hollow account collector
